database: rename heading column in cards to headline

+++ database/schema.sql
@@ -2,7 +2,7 @@
 -- PostgreSQL database dump
 --
 
-\restrict vc4VugPpA7WTa2K0dFM9acjBUhQnvxmtAgrBi34juAOHFPXnniGEzKkl0h1jUHp
+\restrict dGsHx9ipa2gODb2S9cKl6h2eqF1cIWxNxqwFxazD2Qp4zcIwKBbbxjjQBrh6Alf
 
 -- Dumped from database version 18.0
 -- Dumped by pg_dump version 18.0
@@ -90,8 +90,8 @@ ALTER TYPE flashback.expertise_level OWNER TO flashback;
 
 CREATE TYPE flashback.network_activity AS ENUM (
     'signup',
-    'login',
-    'logout',
+    'signin',
+    'signout',
     'upload',
     'download'
 );
@@ -163,22 +163,6 @@ CREATE TYPE flashback.user_action AS ENUM (
 
 
 ALTER TYPE flashback.user_action OWNER TO flashback;
-
---
--- Name: user_progress; Type: TYPE; Schema: flashback; Owner: flashback
---
-
-CREATE TYPE flashback.user_progress AS ENUM (
-    'reading',
-    'paused',
-    'abandoned',
-    'annotated',
-    'mastered',
-    'completed'
-);
-
-
-ALTER TYPE flashback.user_progress OWNER TO flashback;
 
 --
 -- Name: user_state; Type: TYPE; Schema: flashback; Owner: flashback
@@ -429,18 +413,18 @@ ALTER PROCEDURE flashback.create_block(IN card integer, IN type flashback.conten
 -- Name: create_card(text); Type: FUNCTION; Schema: flashback; Owner: flashback
 --
 
-CREATE FUNCTION flashback.create_card(heading text) RETURNS integer
+CREATE FUNCTION flashback.create_card(headline text) RETURNS integer
     LANGUAGE plpgsql
     AS $$
 declare card integer;
 begin
-    insert into cards (heading) values (heading) returning id into card;
+    insert into cards (headline) values (headline) returning id into card;
     return card;
 end;
 $$;
 
 
-ALTER FUNCTION flashback.create_card(heading text) OWNER TO flashback;
+ALTER FUNCTION flashback.create_card(headline text) OWNER TO flashback;
 
 --
 -- Name: create_resource(character varying, flashback.resource_type, flashback.section_pattern, character varying, character varying, character varying); Type: FUNCTION; Schema: flashback; Owner: flashback
@@ -673,19 +657,19 @@ CREATE PROCEDURE flashback.edit_block(IN card integer, IN block integer, IN cont
 ALTER PROCEDURE flashback.edit_block(IN card integer, IN block integer, IN content text) OWNER TO flashback;
 
 --
--- Name: edit_card_heading(integer, character varying); Type: PROCEDURE; Schema: flashback; Owner: flashback
+-- Name: edit_card_headline(integer, character varying); Type: PROCEDURE; Schema: flashback; Owner: flashback
 --
 
-CREATE PROCEDURE flashback.edit_card_heading(IN card integer, IN new_heading character varying)
+CREATE PROCEDURE flashback.edit_card_headline(IN card integer, IN new_headline character varying)
     LANGUAGE plpgsql
     AS $$
 begin
-    update cards set heading = new_heading where id = card;
+    update cards set headline = new_headline where id = card;
 end;
 $$;
 
 
-ALTER PROCEDURE flashback.edit_card_heading(IN card integer, IN new_heading character varying) OWNER TO flashback;
+ALTER PROCEDURE flashback.edit_card_headline(IN card integer, IN new_headline character varying) OWNER TO flashback;
 
 --
 -- Name: edit_resource_name(integer, character varying); Type: PROCEDURE; Schema: flashback; Owner: flashback
@@ -774,13 +758,13 @@ CREATE FUNCTION flashback.estimate_read_time(card_id integer) RETURNS integer
     LANGUAGE plpgsql
     AS $$
 declare time_per_word float = 0.4;
-declare heading_multiplier integer = 2;
+declare headline_multiplier integer = 2;
 declare text_multiplier integer = 1;
 declare code_multiplier integer = 5;
 declare headline_weight integer = 0;
 declare content_weight integer = 0;
 begin
-    select array_length(tsvector_to_array(to_tsvector('simple', heading)), 1) * time_per_word * heading_multiplier into headline_weight from cards where id = card_id;
+    select array_length(tsvector_to_array(to_tsvector('simple', headline)), 1) * time_per_word * headline_multiplier into headline_weight from cards where id = card_id;
     select sum(array_length(tsvector_to_array(to_tsvector('simple', content)), 1)) * case when type = 'text' then text_multiplier when type = 'code' then code_multiplier end into content_weight from blocks where card = card_id group by content, type;
     return headline_weight + content_weight;
 end; $$;
@@ -906,22 +890,22 @@ ALTER FUNCTION flashback.get_cards(user_id integer, subject_id integer, max_leve
 -- Name: get_duplicate_card(integer); Type: FUNCTION; Schema: flashback; Owner: flashback
 --
 
-CREATE FUNCTION flashback.get_duplicate_card(card_id integer) RETURNS TABLE(rid integer, sid integer, "position" integer, card integer, resource character varying, section character varying, state flashback.card_state, heading text)
+CREATE FUNCTION flashback.get_duplicate_card(card_id integer) RETURNS TABLE(rid integer, sid integer, "position" integer, card integer, resource character varying, section character varying, state flashback.card_state, headline text)
     LANGUAGE plpgsql
     AS $$
-declare card_heading text;
+declare card_headline text;
 begin
-    select c.heading into card_heading
+    select c.headline into card_headline
     from cards c
     where c.id = card_id;
 
     return query
-    select sc.resource as rid, sc.section as sid, sc.position, sc.card, r.name as resource, se.name as section, c.state, c.heading
+    select sc.resource as rid, sc.section as sid, sc.position, sc.card, r.name as resource, se.name as section, c.state, c.headline
     from cards c
     join sections_cards sc on c.id = sc.card
     join sections se on se.resource = sc.resource and se.position = sc.section
     join resources r on r.id = sc.resource
-    where c.heading = card_heading and sc.card <> card_id
+    where c.headline = card_headline and sc.card <> card_id
     order by sc.resource, sc.section, sc.position;
 end;
 $$;
@@ -933,12 +917,12 @@ ALTER FUNCTION flashback.get_duplicate_card(card_id integer) OWNER TO flashback;
 -- Name: get_lost_cards(); Type: FUNCTION; Schema: flashback; Owner: flashback
 --
 
-CREATE FUNCTION flashback.get_lost_cards() RETURNS TABLE(sid integer, level flashback.expertise_level, tid integer, "position" integer, card integer, subject character varying, topic character varying, state flashback.card_state, heading text)
+CREATE FUNCTION flashback.get_lost_cards() RETURNS TABLE(sid integer, level flashback.expertise_level, tid integer, "position" integer, card integer, subject character varying, topic character varying, state flashback.card_state, headline text)
     LANGUAGE plpgsql
     AS $$
 begin
     return query
-    select tc.subject as sid, tc.level, tc.topic as tid, tc.position, tc.card, s.name as subject, t.name as topic, c.state, c.heading
+    select tc.subject as sid, tc.level, tc.topic as tid, tc.position, tc.card, s.name as subject, t.name as topic, c.state, c.headline
     from topics_cards tc
     join topics t on t.subject = tc.subject and t.position = tc.topic and t.level = tc.level
     join subjects s on s.id = tc.subject
@@ -1141,9 +1125,16 @@ ALTER FUNCTION flashback.get_sections(resource integer) OWNER TO flashback;
 -- Name: get_sections_cards(integer, integer); Type: FUNCTION; Schema: flashback; Owner: flashback
 --
 
-CREATE FUNCTION flashback.get_sections_cards(resource integer, section integer) RETURNS TABLE(id integer, "position" integer, state flashback.card_state, heading text)
+CREATE FUNCTION flashback.get_sections_cards(resource integer, section integer) RETURNS TABLE(id integer, "position" integer, state flashback.card_state, headline text)
     LANGUAGE plpgsql
-    AS $$ begin return query select cards.id, sc."position", cards.state, cards.heading from sections_cards sc join cards on cards.id = sc.card where sc.resource = get_sections_cards.resource and sc.section = get_sections_cards.section; end; $$;
+    AS $$
+begin
+    return query
+    select cards.id, sc."position", cards.state, cards.headline
+    from sections_cards sc
+    join cards on cards.id = sc.card
+    where sc.resource = get_sections_cards.resource and sc.section = get_sections_cards.section;
+end; $$;
 
 
 ALTER FUNCTION flashback.get_sections_cards(resource integer, section integer) OWNER TO flashback;
@@ -1235,12 +1226,12 @@ ALTER FUNCTION flashback.get_topics(roadmap integer, milestone integer) OWNER TO
 -- Name: get_topics_cards(integer, integer, integer); Type: FUNCTION; Schema: flashback; Owner: flashback
 --
 
-CREATE FUNCTION flashback.get_topics_cards(roadmap_id integer, subject_id integer, topic_position integer) RETURNS TABLE(card integer, "position" integer, state flashback.card_state, heading text)
+CREATE FUNCTION flashback.get_topics_cards(roadmap_id integer, subject_id integer, topic_position integer) RETURNS TABLE(card integer, "position" integer, state flashback.card_state, headline text)
     LANGUAGE plpgsql
     AS $$
 begin
     return query
-    select cards.id, tc."position", cards.state, cards.heading
+    select cards.id, tc."position", cards.state, cards.headline
     from milestones m
     join topics_cards tc on tc.subject = m.subject
     join cards on cards.id = tc.card
@@ -1257,12 +1248,12 @@ ALTER FUNCTION flashback.get_topics_cards(roadmap_id integer, subject_id integer
 -- Name: get_unreviewed_sections_cards(); Type: FUNCTION; Schema: flashback; Owner: flashback
 --
 
-CREATE FUNCTION flashback.get_unreviewed_sections_cards() RETURNS TABLE(rid integer, sid integer, "position" integer, card integer, resource character varying, section character varying, state flashback.card_state, heading text)
+CREATE FUNCTION flashback.get_unreviewed_sections_cards() RETURNS TABLE(rid integer, sid integer, "position" integer, card integer, resource character varying, section character varying, state flashback.card_state, headline text)
     LANGUAGE plpgsql
     AS $$
 begin
     return query
-    select sc.resource, sc.section, sc.position, sc.card, r.name, s.name, c.state, c.heading
+    select sc.resource, sc.section, sc.position, sc.card, r.name, s.name, c.state, c.headline
     from sections_cards sc
     join sections s on s.resource = sc.resource and s.position = sc.section
     join resources r on r.id = sc.resource
@@ -1279,12 +1270,12 @@ ALTER FUNCTION flashback.get_unreviewed_sections_cards() OWNER TO flashback;
 -- Name: get_unreviewed_topics_cards(); Type: FUNCTION; Schema: flashback; Owner: flashback
 --
 
-CREATE FUNCTION flashback.get_unreviewed_topics_cards() RETURNS TABLE(sid integer, level flashback.expertise_level, tid integer, "position" integer, card integer, subject character varying, topic character varying, state flashback.card_state, heading text)
+CREATE FUNCTION flashback.get_unreviewed_topics_cards() RETURNS TABLE(sid integer, level flashback.expertise_level, tid integer, "position" integer, card integer, subject character varying, topic character varying, state flashback.card_state, headline text)
     LANGUAGE plpgsql
     AS $$
 begin
     return query
-    select tc.subject as sid, tc.level, tc.topic as tid, tc.position, tc.card, s.name as subject, t.name as topic, c.state, c.heading
+    select tc.subject as sid, tc.level, tc.topic as tid, tc.position, tc.card, s.name as subject, t.name as topic, c.state, c.headline
     from topics_cards tc
     join topics t on t.subject = tc.subject and t.position = tc.topic and t.level = tc.level
     join subjects s on s.id = tc.subject
@@ -1307,6 +1298,23 @@ CREATE FUNCTION flashback.get_unshelved_resources() RETURNS TABLE(resource integ
 
 
 ALTER FUNCTION flashback.get_unshelved_resources() OWNER TO flashback;
+
+--
+-- Name: get_user(character varying); Type: FUNCTION; Schema: flashback; Owner: flashback
+--
+
+CREATE FUNCTION flashback.get_user(user_email character varying) RETURNS TABLE(id integer, name character varying, email character varying, hash character varying, state flashback.user_state, verified boolean, joined timestamp with time zone, token character varying, device character varying)
+    LANGUAGE plpgsql
+    AS $$
+begin
+    return query
+    select u.id, u.name, u.email, u.hash, u.state, u.verified, u.joined, null::character varying, null::character varying
+    from users u
+    where u.email = user_email;
+end; $$;
+
+
+ALTER FUNCTION flashback.get_user(user_email character varying) OWNER TO flashback;
 
 --
 -- Name: get_user(integer, character varying); Type: FUNCTION; Schema: flashback; Owner: flashback
@@ -1516,7 +1524,7 @@ ALTER PROCEDURE flashback.merge_blocks(IN selected_card integer, VARIADIC block_
 -- Name: merge_cards(integer, integer, character varying); Type: PROCEDURE; Schema: flashback; Owner: flashback
 --
 
-CREATE PROCEDURE flashback.merge_cards(IN lhs integer, IN rhs integer, IN new_heading character varying)
+CREATE PROCEDURE flashback.merge_cards(IN lhs integer, IN rhs integer, IN new_headline character varying)
     LANGUAGE plpgsql
     AS $$
 declare rhs_top_position integer;
@@ -1525,7 +1533,7 @@ begin
 
     update blocks set position = position + rhs_top_position, card = lhs where card = rhs;
 
-    update cards set heading = new_heading where id = lhs;
+    update cards set headline = new_headline where id = lhs;
 
     if not exists (
         select 1 from topics_cards tc join topics_cards tcc on tcc.subject = tc.subject and tcc.topic = tc.topic and tcc.level = tc.level where tc.card = lhs and tcc.card = rhs
@@ -1546,7 +1554,7 @@ end;
 $$;
 
 
-ALTER PROCEDURE flashback.merge_cards(IN lhs integer, IN rhs integer, IN new_heading character varying) OWNER TO flashback;
+ALTER PROCEDURE flashback.merge_cards(IN lhs integer, IN rhs integer, IN new_headline character varying) OWNER TO flashback;
 
 --
 -- Name: move_card_to_section(integer, integer, integer, integer); Type: FUNCTION; Schema: flashback; Owner: flashback
@@ -1767,6 +1775,20 @@ end; $$;
 ALTER PROCEDURE flashback.reset_password(IN user_id integer, IN new_hash character varying) OWNER TO flashback;
 
 --
+-- Name: revoke_session(integer, character varying); Type: PROCEDURE; Schema: flashback; Owner: flashback
+--
+
+CREATE PROCEDURE flashback.revoke_session(IN user_id integer, IN active_token character varying)
+    LANGUAGE plpgsql
+    AS $$
+begin
+    delete from sessions where "user" = user_id and token = active_token;
+end; $$;
+
+
+ALTER PROCEDURE flashback.revoke_session(IN user_id integer, IN active_token character varying) OWNER TO flashback;
+
+--
 -- Name: revoke_sessions_except(integer, character varying); Type: PROCEDURE; Schema: flashback; Owner: flashback
 --
 
@@ -1774,7 +1796,9 @@ CREATE PROCEDURE flashback.revoke_sessions_except(IN user_id integer, IN active_
     LANGUAGE plpgsql
     AS $$
 begin
-    delete from sessions where "user" = user_id and token <> active_token;
+    if exists (select device from sessions where "user" = user_id and token = active_token) then
+        delete from sessions where "user" = user_id and token <> active_token;
+    end if;
 end; $$;
 
 
@@ -1944,23 +1968,12 @@ ALTER TABLE flashback.blocks_activities ALTER COLUMN id ADD GENERATED ALWAYS AS 
 
 
 --
--- Name: card_position; Type: TABLE; Schema: flashback; Owner: flashback
---
-
-CREATE TABLE flashback.card_position (
-    "coalesce" integer
-);
-
-
-ALTER TABLE flashback.card_position OWNER TO flashback;
-
---
 -- Name: cards; Type: TABLE; Schema: flashback; Owner: flashback
 --
 
 CREATE TABLE flashback.cards (
     id integer NOT NULL,
-    heading text,
+    headline text,
     state flashback.card_state DEFAULT 'draft'::flashback.card_state NOT NULL
 );
 
@@ -2053,17 +2066,6 @@ ALTER TABLE flashback.milestones_activities ALTER COLUMN id ADD GENERATED ALWAYS
     CACHE 1
 );
 
-
---
--- Name: most_recent; Type: TABLE; Schema: flashback; Owner: flashback
---
-
-CREATE TABLE flashback.most_recent (
-    "?column?" interval
-);
-
-
-ALTER TABLE flashback.most_recent OWNER TO flashback;
 
 --
 -- Name: network_activities; Type: TABLE; Schema: flashback; Owner: flashback
@@ -2252,24 +2254,14 @@ ALTER TABLE flashback.roadmaps ALTER COLUMN id ADD GENERATED ALWAYS AS IDENTITY 
 
 
 --
--- Name: section_name; Type: TABLE; Schema: flashback; Owner: flashback
---
-
-CREATE TABLE flashback.section_name (
-    pattern flashback.section_pattern
-);
-
-
-ALTER TABLE flashback.section_name OWNER TO flashback;
-
---
 -- Name: sections; Type: TABLE; Schema: flashback; Owner: flashback
 --
 
 CREATE TABLE flashback.sections (
     resource integer NOT NULL,
     "position" integer NOT NULL,
-    name character varying(200)
+    name character varying(200),
+    link character varying(2000)
 );
 
 
@@ -2344,6 +2336,36 @@ CREATE TABLE flashback.shelves (
 
 
 ALTER TABLE flashback.shelves OWNER TO flashback;
+
+--
+-- Name: shelves_activities; Type: TABLE; Schema: flashback; Owner: flashback
+--
+
+CREATE TABLE flashback.shelves_activities (
+    id integer NOT NULL,
+    "user" integer NOT NULL,
+    resource integer NOT NULL,
+    subject integer NOT NULL,
+    action flashback.user_action,
+    "time" timestamp with time zone
+);
+
+
+ALTER TABLE flashback.shelves_activities OWNER TO flashback;
+
+--
+-- Name: shelves_activities_id_seq; Type: SEQUENCE; Schema: flashback; Owner: flashback
+--
+
+ALTER TABLE flashback.shelves_activities ALTER COLUMN id ADD GENERATED ALWAYS AS IDENTITY (
+    SEQUENCE NAME flashback.shelves_activities_id_seq
+    START WITH 1
+    INCREMENT BY 1
+    NO MINVALUE
+    NO MAXVALUE
+    CACHE 1
+);
+
 
 --
 -- Name: subjects; Type: TABLE; Schema: flashback; Owner: flashback
@@ -2648,6 +2670,14 @@ ALTER TABLE ONLY flashback.sessions
 
 
 --
+-- Name: shelves_activities shelves_activities_pkey; Type: CONSTRAINT; Schema: flashback; Owner: flashback
+--
+
+ALTER TABLE ONLY flashback.shelves_activities
+    ADD CONSTRAINT shelves_activities_pkey PRIMARY KEY (id);
+
+
+--
 -- Name: shelves shelves_pkey; Type: CONSTRAINT; Schema: flashback; Owner: flashback
 --
 
@@ -2804,7 +2834,7 @@ ALTER TABLE ONLY flashback.cards_activities
 --
 
 ALTER TABLE ONLY flashback.milestones_activities
-    ADD CONSTRAINT milestones_activities_roadmap_subject_fkey FOREIGN KEY (roadmap, subject) REFERENCES flashback.milestones(roadmap, subject) ON UPDATE CASCADE ON DELETE SET NULL;
+    ADD CONSTRAINT milestones_activities_roadmap_subject_fkey FOREIGN KEY (roadmap, subject) REFERENCES flashback.milestones(roadmap, subject) ON UPDATE CASCADE ON DELETE CASCADE;
 
 
 --
@@ -2944,6 +2974,30 @@ ALTER TABLE ONLY flashback.sessions
 
 
 --
+-- Name: shelves_activities shelves_activities_resource_fkey; Type: FK CONSTRAINT; Schema: flashback; Owner: flashback
+--
+
+ALTER TABLE ONLY flashback.shelves_activities
+    ADD CONSTRAINT shelves_activities_resource_fkey FOREIGN KEY (resource) REFERENCES flashback.resources(id) ON UPDATE CASCADE ON DELETE CASCADE;
+
+
+--
+-- Name: shelves_activities shelves_activities_subject_fkey; Type: FK CONSTRAINT; Schema: flashback; Owner: flashback
+--
+
+ALTER TABLE ONLY flashback.shelves_activities
+    ADD CONSTRAINT shelves_activities_subject_fkey FOREIGN KEY (subject) REFERENCES flashback.subjects(id) ON UPDATE CASCADE ON DELETE CASCADE;
+
+
+--
+-- Name: shelves_activities shelves_activities_user_fkey; Type: FK CONSTRAINT; Schema: flashback; Owner: flashback
+--
+
+ALTER TABLE ONLY flashback.shelves_activities
+    ADD CONSTRAINT shelves_activities_user_fkey FOREIGN KEY ("user") REFERENCES flashback.users(id) ON UPDATE CASCADE ON DELETE CASCADE;
+
+
+--
 -- Name: shelves shelves_resource_fkey; Type: FK CONSTRAINT; Schema: flashback; Owner: flashback
 --
 
@@ -3027,5 +3081,5 @@ ALTER TABLE ONLY flashback.users_roadmaps
 -- PostgreSQL database dump complete
 --
 
-\unrestrict vc4VugPpA7WTa2K0dFM9acjBUhQnvxmtAgrBi34juAOHFPXnniGEzKkl0h1jUHp
+\unrestrict dGsHx9ipa2gODb2S9cKl6h2eqF1cIWxNxqwFxazD2Qp4zcIwKBbbxjjQBrh6Alf
